+++ include/MCL/DefaultBuilders.hpp
@@ -45,10 +45,10 @@ typedef std::function<std::shared_ptr<Material> ( std::string type, std::vector<
 
 
 static void trimesh_copy( std::shared_ptr<mcl::TriangleMesh> to_mesh, trimesh::TriMesh *from_mesh ){
-	for( int i=0; i<from_mesh->vertices.size(); ++i ){ to_mesh->vertices.push_back( mcl::Vec3d( from_mesh->vertices[i][0], from_mesh->vertices[i][1], from_mesh->vertices[i][2] ) ); }
-	for( int i=0; i<from_mesh->colors.size(); ++i ){ to_mesh->colors.push_back( mcl::Vec3d( from_mesh->colors[i][0], from_mesh->colors[i][1], from_mesh->colors[i][2] ) ); }
+	for( int i=0; i<from_mesh->vertices.size(); ++i ){ to_mesh->vertices.push_back( mcl::Vec3f( from_mesh->vertices[i][0], from_mesh->vertices[i][1], from_mesh->vertices[i][2] ) ); }
+	for( int i=0; i<from_mesh->colors.size(); ++i ){ to_mesh->colors.push_back( mcl::Vec3f( from_mesh->colors[i][0], from_mesh->colors[i][1], from_mesh->colors[i][2] ) ); }
 	for( int i=0; i<from_mesh->faces.size(); ++i ){ to_mesh->faces.push_back( mcl::Vec3i( from_mesh->faces[i][0], from_mesh->faces[i][1], from_mesh->faces[i][2] ) ); }
-	for( int i=0; i<from_mesh->texcoords.size(); ++i ){ to_mesh->texcoords.push_back( mcl::Vec2d( from_mesh->texcoords[i][0], from_mesh->texcoords[i][1] ) ); }
+	for( int i=0; i<from_mesh->texcoords.size(); ++i ){ to_mesh->texcoords.push_back( mcl::Vec2f( from_mesh->texcoords[i][0], from_mesh->texcoords[i][1] ) ); }
 	to_mesh->update();
 }
